refactor: 날짜 및 닉네임 확인 메서드 추가
