+++ src/main/java/attendance/service/AttendanceService.java
@@ -1,16 +1,14 @@
 package attendance.service;
 
-import attendance.AttendanceRecord;
-import attendance.AttendanceRepository;
-import attendance.domain.*;
-
+import attendance.repository.AttendanceRepository;
+import attendance.Constants;
+import attendance.domain.enums.*;
+import attendance.domain.model.*;
+import attendance.domain.analysis.*;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.Set;
-import java.util.HashSet;
 
 public class AttendanceService {
     private final AttendanceRepository repository;
@@ -19,18 +17,29 @@ public class AttendanceService {
         this.repository = repository;
     }
 
-    public Attendance checkAttendance(Nickname nickname, AttendanceTime time) {
-        LocalDate today = LocalDate.of(2024,12,13);
+    public void checkDay(){
+        LocalDate today = Constants.getToday();
         AttendanceDate date = new AttendanceDate(today);
-
         if (!date.isAttendanceDay()) {
             throw new IllegalArgumentException("오늘은 출석일이 아닙니다.");
         }
+    }
+    public void checkNickname(Nickname nickname){
+        List<Nickname> nicknames = repository.findAllNicknames();
+        boolean exists = nicknames.stream()
+                .anyMatch(n -> n.getValue().equals(nickname.getValue()));
+        if (!exists) {
+            throw new IllegalArgumentException("등록되지 않은 닉네임입니다.");
+        }
+    }
+
+    public Attendance checkAttendance(Nickname nickname, AttendanceTime time) {
+        LocalDate today = Constants.getToday();
+        AttendanceDate date = new AttendanceDate(today);
 
         if (repository.existsByNicknameAndDate(nickname, date)) {
             throw new IllegalArgumentException("이미 출석한 기록이 있습니다. 수정 기능을 이용해주세요.");
         }
-
         Attendance attendance = new Attendance(nickname, date, time);
         repository.save(attendance);
         return attendance;
@@ -41,20 +50,12 @@ public class AttendanceService {
             throw new IllegalArgumentException("해당 날짜에 출석 기록이 없습니다.");
         }
 
-        List<Attendance> attendances = repository.findAllByNickname(nickname);
-        Attendance oldAttendance = attendances.stream()
-                .filter(a -> a.getDate().getValue().equals(date.getValue()))
-                .findFirst()
-                .orElseThrow();
-
         repository.update(nickname, date, time);
-        //정리 필요
-        Attendance newAttendance = new Attendance(nickname, date, time);
-        return newAttendance;
+        return new Attendance(nickname, date, time);
     }
 
     public Attendance getOldAttendance(Nickname nickname, AttendanceDate date){
-        return this.repository.findByNicknameAndDate(nickname,date);
+        return repository.findByNicknameAndDate(nickname, date);
     }
 
     public AttendanceRecord getAttendanceRecord(Nickname nickname) {
@@ -64,31 +65,23 @@ public class AttendanceService {
 
     public List<RiskCrew> getRiskCrews() {
         List<RiskCrew> riskCrews = new ArrayList<>();
-        // 중복 방지를 위한 Set 사용
-        Set<String> processedNicknames = new HashSet<>();
-
         List<Nickname> nicknames = repository.findAllNicknames();
 
         for (Nickname nickname : nicknames) {
-            String nicknameValue = nickname.getValue();
+            try {
+                AttendanceRecord record = getAttendanceRecord(nickname);
+                DisciplinaryStatus status = record.getDisciplinaryStatus();
 
-            // 이미 처리한 닉네임은 건너뛰기
-            if (processedNicknames.contains(nicknameValue)) {
-                continue;
-            }
-
-            processedNicknames.add(nicknameValue);
-
-            AttendanceRecord record = getAttendanceRecord(nickname);
-            DisciplinaryStatus status = record.getDisciplinaryStatus();
-
-            if (status != DisciplinaryStatus.NORMAL) {
-                riskCrews.add(new RiskCrew(
-                        nickname,
-                        record.getAbsentCount(),
-                        record.getLateCount(),
-                        status
-                ));
+                if (status != DisciplinaryStatus.NORMAL) {
+                    riskCrews.add(new RiskCrew(
+                            nickname,
+                            record.getAbsentCount(),
+                            record.getLateCount(),
+                            status
+                    ));
+                }
+            } catch (Exception e) {
+                System.out.println("Debug: " + nickname.getValue() + " 처리 중 오류: " + e.getMessage());
             }
         }
